requirements/base.txt
Django==2.2.4
django-active-link==0.1.5
django-picklefield==2.0
django-widget-tweaks==1.4.3
nltk==3.4.5
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
idna==2.8
language-check==1.1
oauthlib==3.0.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
singledispatch==3.4.0.3
six==1.12.0
textblob==0.15.2
urllib3==1.24.2
